# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/events": {
      "post": {
        "operationId": "EventsController_create",
        "summary": "",
        "description": "Creates an event from the Dto object.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateEventDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EventDocument" }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          }
        },
        "tags": ["Events"],
        "security": [{ "bearer": [] }]
      },
      "get": {
        "operationId": "EventsController_find",
        "summary": "",
        "description": "Gets all event documents from the database which satisfy the filter conditions.",
        "parameters": [
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "The name of the event.",
            "example": "American Gods",
            "schema": { "type": "string" }
          },
          {
            "name": "bookSearchQuery",
            "required": false,
            "in": "query",
            "description": "The query string which will search among book titles.",
            "example": "American",
            "schema": { "type": "string" }
          },
          {
            "name": "bookIds",
            "required": false,
            "in": "query",
            "description": "A list of object IDs pointing to book documents.",
            "schema": { "type": "array", "items": { "type": "string" } }
          },
          {
            "name": "threads",
            "required": false,
            "in": "query",
            "description": "A list of channel/thread IDs.",
            "schema": { "type": "array", "items": { "type": "string" } }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "The status of the event.",
            "example": "Approved",
            "schema": {
              "enum": [
                "Requested",
                "Approved",
                "Announced",
                "Ongoing",
                "Completed",
                "Rejected",
                "Cancelled"
              ],
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "The type of the event.",
            "example": "BuddyRead",
            "schema": {
              "enum": [
                "BuddyRead",
                "MonthlyRead",
                "ShortStoryRead",
                "PoetryRead",
                "Other"
              ],
              "type": "string"
            }
          },
          {
            "name": "startDateBefore",
            "required": false,
            "in": "query",
            "description": "The start date of the event should be before this date.",
            "schema": { "format": "date-time", "type": "string" }
          },
          {
            "name": "startDateAfter",
            "required": false,
            "in": "query",
            "description": "The start date of the event should be after this date.",
            "schema": { "format": "date-time", "type": "string" }
          },
          {
            "name": "endDateBefore",
            "required": false,
            "in": "query",
            "description": "The end date of the event should be before this date.",
            "schema": { "format": "date-time", "type": "string" }
          },
          {
            "name": "endDateAfter",
            "required": false,
            "in": "query",
            "description": "The end date of the event should be after this date.",
            "schema": { "format": "date-time", "type": "string" }
          },
          {
            "name": "participantIds",
            "required": false,
            "in": "query",
            "description": "The ID of the participant of the event should be in this list.\nThis covers requestors, interested, readers etc all participants.",
            "schema": { "type": "array", "items": { "type": "string" } }
          },
          {
            "name": "requestedByIds",
            "required": false,
            "in": "query",
            "description": "The ID of the requestor of the event should be in this list.",
            "schema": { "type": "array", "items": { "type": "string" } }
          },
          {
            "name": "interestedIds",
            "required": false,
            "in": "query",
            "description": "The ID of the interested people of the event should be in this list.",
            "schema": { "type": "array", "items": { "type": "string" } }
          },
          {
            "name": "readerIds",
            "required": false,
            "in": "query",
            "description": "The ID of the reader of the event should be in this list.",
            "schema": { "type": "array", "items": { "type": "string" } }
          },
          {
            "name": "leaderIds",
            "required": false,
            "in": "query",
            "description": "The ID of the leader of the event should be in this list.",
            "schema": { "type": "array", "items": { "type": "string" } }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/EventDocument" }
                }
              }
            }
          }
        },
        "tags": ["Events"],
        "security": [{ "bearer": [] }]
      }
    },
    "/api/events/createFromUrl": {
      "post": {
        "operationId": "EventsController_createFromUrl",
        "summary": "",
        "description": "Creates an event from the book URL and Dto object.",
        "parameters": [
          {
            "name": "url",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateEventDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EventDocument" }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          }
        },
        "tags": ["Events"],
        "security": [{ "bearer": [] }]
      }
    },
    "/api/events/{id}": {
      "get": {
        "operationId": "EventsController_findOne",
        "summary": "",
        "description": "Gets the event document with the given ID from the database.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EventDocument" }
              }
            }
          }
        },
        "tags": ["Events"],
        "security": [{ "bearer": [] }]
      },
      "patch": {
        "operationId": "EventsController_update",
        "summary": "",
        "description": "Updates the event with the given ID and Dto in the database.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateEventDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EventDocument" }
              }
            }
          }
        },
        "tags": ["Events"],
        "security": [{ "bearer": [] }]
      },
      "delete": {
        "operationId": "EventsController_remove",
        "summary": "",
        "description": "Deletes the event document from the DB.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "type": "boolean" } }
            }
          }
        },
        "tags": ["Events"],
        "security": [{ "bearer": [] }]
      }
    },
...
  },
  "info": {
    "title": "OWS",
    "description": "API description for the Organized Web Server",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
      "AuthorDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the author.",
            "example": "Neil Gaiman"
          },
          "url": {
            "type": "string",
            "description": "The GR/SG link to that author's page.",
            "example": "https://www.goodreads.com/author/show/1221698.Neil_Gaiman"
          }
        },
        "required": ["name", "url"]
      },
...
      "BookDocument": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "The title of the book.",
            "example": "American Gods"
          },
          "authors": {
            "description": "A list of Author DTOs.",
            "type": "array",
            "items": { "$ref": "#/components/schemas/AuthorDto" }
          },
          "url": {
            "type": "string",
            "description": "The GR/SG link to the book page.",
            "example": "https://www.goodreads.com/book/show/30165203-american-gods"
          },
          "genres": {
            "description": "The list of genres for the book.",
            "example": ["Fantasy", "Fiction", "Science Fiction"],
            "type": "array",
            "items": { "type": "string" }
          },
          "coverUrl": {
            "type": "string",
            "description": "The link to the cover of the book.",
            "example": "https://images-na.ssl-images-amazon.com/images/S/compressed.photo.goodreads.com/books/1462924585i/30165203.jpg"
          },
          "_id": { "type": "string" }
        },
        "required": ["title", "authors", "url", "genres", "coverUrl", "_id"]
      },
      "DateRange": {
        "type": "object",
        "properties": {
          "startDate": {
            "format": "date-time",
            "type": "string",
            "description": "The start date of the event."
          },
          "endDate": {
            "format": "date-time",
            "type": "string",
            "description": "The end date of the event."
          }
        },
        "required": ["startDate", "endDate"]
      },
      "ParticipantDto": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "description": "The object ID of the user (not the user ID)."
          },
          "points": {
            "type": "number",
            "description": "The number of points to be assigned to the particpant."
          }
        },
        "required": ["user", "points"]
      },
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the event.",
            "example": "American Gods"
          },
          "book": {
            "type": "string",
            "description": "The object ID of the book document."
          },
          "threads": {
            "description": "A list of channel/thread IDs.",
            "type": "array",
            "items": { "type": "string" }
          },
          "status": {
            "description": "The status of the event.",
            "example": "Approved",
            "enum": [
              "Requested",
              "Approved",
              "Announced",
              "Ongoing",
              "Completed",
              "Rejected",
              "Cancelled"
            ],
            "type": "string"
          },
          "type": {
            "description": "The type of the event.",
            "example": "BuddyRead",
            "enum": [
              "BuddyRead",
              "MonthlyRead",
              "ShortStoryRead",
              "PoetryRead",
              "Other"
            ],
            "type": "string"
          },
          "dates": {
            "description": "The dates for the event.",
            "allOf": [{ "$ref": "#/components/schemas/DateRange" }]
          },
          "requestedBy": {
            "description": "The participant who requested the event.",
            "allOf": [{ "$ref": "#/components/schemas/ParticipantDto" }]
          },
          "interested": {
            "description": "A list of participants who show interest in the event.",
            "type": "array",
            "items": { "$ref": "#/components/schemas/ParticipantDto" }
          },
          "readers": {
            "description": "The participants who read the book in the event duration.",
            "type": "array",
            "items": { "$ref": "#/components/schemas/ParticipantDto" }
          },
          "leaders": {
            "description": "The participants who lead the discussion for the event.",
            "type": "array",
            "items": { "$ref": "#/components/schemas/ParticipantDto" }
          },
          "description": {
            "type": "string",
            "description": "The description of the event."
          }
        }
      },
      "ProfileDto": {
        "type": "object",
        "properties": {
          "bio": {
            "type": "string",
            "description": "A short description of the user."
          }
        },
        "required": ["bio"]
      },
      "UserDocument": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "userId": {
            "type": "string",
            "description": "The unique ID for the user."
          },
          "name": { "type": "string", "description": "The username." },
          "joinDate": {
            "format": "date-time",
            "type": "string",
            "description": "The joining date for the user."
          },
          "profile": {
            "description": "The profile information fot the user.",
            "allOf": [{ "$ref": "#/components/schemas/ProfileDto" }]
          }
        },
        "required": ["_id", "userId", "name", "joinDate", "profile"]
      },
      "Participant": {
        "type": "object",
        "properties": {
          "user": {
            "description": "The user object.",
            "allOf": [{ "$ref": "#/components/schemas/UserDocument" }]
          },
          "points": {
            "type": "number",
            "description": "The number of points to be assigned to the particpant."
          }
        },
        "required": ["user", "points"]
      },
      "EventDocument": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "name": {
            "type": "string",
            "description": "The name of the event.",
            "example": "American Gods"
          },
          "book": {
            "description": "The book for which the event is happening.",
            "allOf": [{ "$ref": "#/components/schemas/BookDocument" }]
          },
          "threads": {
            "description": "A list of channel/thread IDs.",
            "type": "array",
            "items": { "type": "string" }
          },
          "status": {
            "description": "The status of the event.",
            "example": "Approved",
            "enum": [
              "Requested",
              "Approved",
              "Announced",
              "Ongoing",
              "Completed",
              "Rejected",
              "Cancelled"
            ],
            "type": "string"
          },
          "type": {
            "description": "The type of the event.",
            "example": "BuddyRead",
            "enum": [
              "BuddyRead",
              "MonthlyRead",
              "ShortStoryRead",
              "PoetryRead",
              "Other"
            ],
            "type": "string"
          },
          "dates": {
            "description": "The dates for the event.",
            "allOf": [{ "$ref": "#/components/schemas/DateRange" }]
          },
          "requestedBy": {
            "description": "The participant who requested the event.",
            "allOf": [{ "$ref": "#/components/schemas/Participant" }]
          },
          "interested": {
            "description": "A list of participants who show interest in the event.",
            "type": "array",
            "items": { "$ref": "#/components/schemas/Participant" }
          },
          "readers": {
            "description": "The participants who read the book in the event duration.",
            "type": "array",
            "items": { "$ref": "#/components/schemas/Participant" }
          },
          "leaders": {
            "description": "The participants who lead the discussion for the event.",
            "type": "array",
            "items": { "$ref": "#/components/schemas/Participant" }
          },
          "description": {
            "type": "string",
            "description": "The description of the event."
          }
        },
        "required": [
          "_id",
          "name",
          "book",
          "threads",
          "status",
          "type",
          "dates",
          "requestedBy",
          "interested",
          "readers",
          "leaders",
          "description"
        ]
      },
      "UpdateEventDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the event.",
            "example": "American Gods"
          },
          "book": {
            "type": "string",
            "description": "The object ID of the book document."
          },
          "threads": {
            "description": "A list of channel/thread IDs.",
            "type": "array",
            "items": { "type": "string" }
          },
          "status": {
            "type": "string",
            "description": "The status of the event.",
            "example": "Approved",
            "enum": [
              "Requested",
              "Approved",
              "Announced",
              "Ongoing",
              "Completed",
              "Rejected",
              "Cancelled"
            ]
          },
          "type": {
            "type": "string",
            "description": "The type of the event.",
            "example": "BuddyRead",
            "enum": [
              "BuddyRead",
              "MonthlyRead",
              "ShortStoryRead",
              "PoetryRead",
              "Other"
            ]
          },
          "dates": {
            "description": "The dates for the event.",
            "allOf": [{ "$ref": "#/components/schemas/DateRange" }]
          },
          "requestedBy": {
            "description": "The participant who requested the event.",
            "allOf": [{ "$ref": "#/components/schemas/ParticipantDto" }]
          },
          "interested": {
            "description": "A list of participants who show interest in the event.",
            "type": "array",
            "items": { "$ref": "#/components/schemas/ParticipantDto" }
          },
          "readers": {
            "description": "The participants who read the book in the event duration.",
            "type": "array",
            "items": { "$ref": "#/components/schemas/ParticipantDto" }
          },
          "leaders": {
            "description": "The participants who lead the discussion for the event.",
            "type": "array",
            "items": { "$ref": "#/components/schemas/ParticipantDto" }
          },
          "description": {
            "type": "string",
            "description": "The description of the event."
          }
        }
      },
...
    }
  }
}